# lineprotocol: optimize decoding